Adding the quiescence part of determining loop invariant computations. I believe that it now is determining all loop invariant computations.
I know that Hongbo wants to use this and I think Raj mentioned that someone else might want to use it as well.
There are more things to check if you want to do hoisting than are included in the present function.
I don't know whether those checks should be a separate function or included in with this existing function.
If anybody has any opinions let me know.