For counting symbols to eliminate unused symbols and gensyms, I changed the way the callback is done. Instead of passing in a callback, pass in an AstWalk entry point (which encodes its own callback) which then gets used.
Change replaceExprWithDict to not use its own traversal routine but to use AstWalk.  This is necessary
because we have to recurse into non-standard AST nodes and the previous traversal wouldn't do that.

The traverse routine became an AstWalk callback routine update_sym that just replaces symbols based on the dictionary.