LivenessAnalysis: refactor to decrease compile time.
This is an attempt to get some typeof() calls out of the liveness pass
and therefore speed up type inference.  I'm not 100% sure the change
is semantics-preserving, but so far, tests seem to pass.

In a quick experiment, this seemed to result in a 0.14s reduction in
compile time for black-scholes (averaged over 10 trial runs).