Refactor various modules to try to improve compile time.
This change eliminates a number of typeof() calls and places where we
were branching on types, in favor of Julia's type-based method
dispatch.

In a quick experiment, this seemed to result in a ~0.5s reduction in
compile time for black-scholes (17.63s vs. 18.13s on my desktop
machine, averaged over 10 trial runs).